stock_market_agent/requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
trustcall
yfinance
numpy
scipy
pandas
python-dotenv
vaderSentiment
Pillow

